+++ tests/unit/test_image.py
@@ -396,6 +396,41 @@ def test_train_local_code(tmpdir, sagemaker_session):
                 assert '%s:/opt/ml/shared' % shared_folder_path in volumes
 
 
+@patch('sagemaker.local.local_session.LocalSession', Mock())
+@patch('sagemaker.local.image._stream_output', Mock())
+@patch('sagemaker.local.image._SageMakerContainer._cleanup', Mock())
+@patch('sagemaker.local.data.get_data_source_instance', Mock())
+@patch('subprocess.Popen', Mock())
+def test_train_local_intermediate_output(tmpdir, sagemaker_session):
+    directories = [str(tmpdir.mkdir('container-root')), str(tmpdir.mkdir('data'))]
+    with patch('sagemaker.local.image._SageMakerContainer._create_tmp_folder',
+               side_effect=directories):
+        instance_count = 2
+        image = 'my-image'
+        sagemaker_container = _SageMakerContainer('local', instance_count, image,
+                                                  sagemaker_session=sagemaker_session)
+
+        output_path = str(tmpdir.mkdir('customer_intermediate_output'))
+        output_data_config = {'S3OutputPath': 'file://%s' % output_path}
+        hyperparameters = {'sagemaker_s3_output': output_path}
+
+        sagemaker_container.train(
+            INPUT_DATA_CONFIG, output_data_config, hyperparameters, TRAINING_JOB_NAME)
+
+        docker_compose_file = os.path.join(sagemaker_container.container_root,
+                                           'docker-compose.yaml')
+        intermediate_folder_path = os.path.join(output_path, 'output/intermediate')
+
+        with open(docker_compose_file, 'r') as f:
+            config = yaml.load(f)
+            assert len(config['services']) == instance_count
+            for h in sagemaker_container.hosts:
+                assert config['services'][h]['image'] == image
+                assert config['services'][h]['command'] == 'train'
+                volumes = config['services'][h]['volumes']
+                assert '%s:/opt/ml/output/intermediate' % intermediate_folder_path in volumes
+
+
 def test_container_has_gpu_support(tmpdir, sagemaker_session):
     instance_count = 1
     image = 'my-image'