+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ CHANGELOG
 
 * doc-fix: Change ``distribution`` to ``distributions``
 * bug-fix: Increase docker-compose http timeout and health check timeout to 120.
+* feature: Local Mode: Add support for intermediate output to a local directory.
 
 1.16.1.post1
 ============